+++ packages/components/src/progress-bar/index.tsx
@@ -20,11 +20,11 @@ function UnforwardedProgressBar(
 	props: WordPressComponentProps< ProgressBarProps, 'progress', false >,
 	ref: ForwardedRef< HTMLProgressElement >
 ) {
-	const { className, value, ...progressProps } = props;
+	const { className, value, width, ...progressProps } = props;
 	const isIndeterminate = ! Number.isFinite( value );
 
 	return (
-		<ProgressBarStyled.Track className={ className }>
+		<ProgressBarStyled.Track className={ className } width={ width }>
 			<ProgressBarStyled.Indicator
 				style={
 					{