+++ packages/components/src/progress-bar/styles.ts
@@ -21,11 +21,12 @@ const animateProgressBar = keyframes( {
 // Width of the indicator for the indeterminate progress bar
 export const INDETERMINATE_TRACK_WIDTH = 50;
 
-export const Track = styled.div`
+export const Track = styled.div< { width?: number } >`
 	position: relative;
 	overflow: hidden;
 	width: 100%;
-	max-width: 160px;
+	max-width: ${ ( { width } ) =>
+		width ? width : CONFIG.progressBarWidth }px;
 	height: ${ CONFIG.borderWidthFocus };
 	/* Text color at 10% opacity */
 	background-color: color-mix(